requirements.txt
apify < 3.0
apify-client
python-dotenv
pydantic >= 2.0
pydantic-ai
requests
httpx
